+++ tests/123_enum_representation.rs
@@ -312,7 +312,7 @@ impl ByteStr {
             type Value = Vec<u8>;
 
             // GRCOV_EXCL_START
-            fn expecting(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+            fn expecting(&self, fmt: &mut core::fmt::Formatter) -> core::fmt::Result {
                 fmt.write_str("a Rusty byte string")
             }
             // GRCOV_EXCL_STOP