+++ tests/465_implicit_some_stack.rs
@@ -9,7 +9,7 @@ fn roundtrip_implicit_some_stack() {
 }
 
 fn check_roundtrip<
-    T: PartialEq + std::fmt::Debug + serde::Serialize + serde::de::DeserializeOwned,
+    T: PartialEq + core::fmt::Debug + serde::Serialize + serde::de::DeserializeOwned,
 >(
     val: T,
     check: &str,