+++ tests/non_identifier_identifier.rs
@@ -10,7 +10,7 @@ macro_rules! test_non_identifier {
                 type Value = FieldName;
 
                 // GRCOV_EXCL_START
-                fn expecting(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+                fn expecting(&self, fmt: &mut core::fmt::Formatter) -> core::fmt::Result {
                     fmt.write_str("an error")
                 }
                 // GRCOV_EXCL_STOP
@@ -32,7 +32,7 @@ macro_rules! test_non_identifier {
                 type Value = Struct;
 
                 // GRCOV_EXCL_START
-                fn expecting(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+                fn expecting(&self, fmt: &mut core::fmt::Formatter) -> core::fmt::Result {
                     fmt.write_str("a struct")
                 }
                 // GRCOV_EXCL_STOP